src/requirements.txt
autopep8
debugpy
fastapi
firebase_admin
flake8
google-cloud-storage
isort
mypy
pytest
python-multipart
uvicorn